imapproc: fix separator/prefix detection
If the root folder is not the first folder to be returned by
the LIST command, prefix detection could go haywire. This
commit addresses the issue by using the first folder argument
as reference argument for the LIST command. It also fixes a
bug which caused '/' to be selected as folder separator by
mistake.