IMAP: Fix issues around cache
If we're only have an in-memory cache, then cache_file would be None
(and therefore we shouldn't try to open it for reading or writing).
Also, in that case, header_cache[folder] might exist but be empty; so
when we get the uidvalidity, we need to prepare for it not existing.