Add SHA1 fingerprints for maildir emails.
This feature adds a fingerprint() method to
mailprocessing.mail.maildir.MaildirMail. The purpose of this
method is providing a canonical message comparison metric for
deduplication and archiving. Please note that this feature is
not available for mailprocessing.mail.imap.ImapMail since that
class only stores the message's headers, not the entire body.

It computes a SHA1 checksum over a selection of headers plus a
plain text representation of the raw message body. By default
it uses the following headers. By default it uses the
following headers:

  From:
  To:
  Subject:
  Date:
  CC:
  Content-Type:

This set of headers should - absent a really screwed up MTA -
be transmission invariant, i.e. they should not change after
the message has passed through a series of MTAs and has
reached its destination. It is possible to supply a custom
list of headers (or None for "all headers") through the
`headers` keyword argument.